Use float math functions instead of casting from double

+++ src/Hazard.cpp
@@ -173,8 +173,8 @@ void Hazard::setAngle(const float& _angle) {
 	while (angle >= static_cast<float>(M_PI)*2) angle -= static_cast<float>(M_PI)*2;
 	while (angle < 0.0) angle += static_cast<float>(M_PI)*2;
 
-	speed.x = base_speed * static_cast<float>(cos(angle));
-	speed.y = base_speed * static_cast<float>(sin(angle));
+	speed.x = base_speed * cosf(angle);
+	speed.y = base_speed * sinf(angle);
 
 	if (directional)
 		animationKind = calcDirection(pos.x, pos.y, pos.x + speed.x, pos.y + speed.y);
